Add new tables to default db
Make sure that the defaultCategories are enabled by default and that the
items category is disabled.

Need to add a table for currency and item icons.

Also add a table for quest information for a future release where the
addOn handles dailies and weeklies by itself without the need of tracking
every quest.